@@ -20,7 +20,7 @@
 	],
 
 	"require": {
-		"php": ">=5.3.2",
+		"php": ">=5.6.0",
 		"behat/mink": "~1.6",
 		"mindplay/annotations": "~1.2",
 		"pimple/pimple": "^2.0|^3.0"
@@ -29,7 +29,7 @@
 	"require-dev": {
 		"aik099/coding-standard": "dev-master",
 		"behat/mink-selenium2-driver": "~1.2",
-		"mockery/mockery": "~0.9",
+		"mockery/mockery": "~1.1",
 		"symfony/phpunit-bridge": "^3.1"
 	},
 
